prototype/requirements.txt
# MobileNet image feature extraction prototype - Dependencies
# Python 3.8+

# Deep learning framework
tensorflow>=2.10.0,<3.0.0

# Image processing
opencv-python>=4.6.0
Pillow>=9.0.0
scikit-image>=0.19.0

# Data processing
numpy>=1.21.0
pandas>=1.3.0

# Visualization
matplotlib>=3.5.0

# Other tools
tqdm>=4.64.0

# HTTP client (for Ollama API)
requests>=2.28.0

# Phase 3 (Hugging Face text generation)
transformers>=4.44.0
torch>=2.1.0
accelerate>=0.30.0

# Phase 5 Web UI (FastAPI)
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
jinja2>=3.1.0
